deleted llm folder and switched to app folder
