--- NLP_LLM/checkVenv.py
@@ -1,9 +1,0 @@
-import sys
-
-def is_venv():
-    return (hasattr(sys, 'real_prefix') or
-            (hasattr(sys, 'base_prefix') and sys.base_prefix != sys.prefix))
-if is_venv():
-    print('inside virtualenv or venv')
-else:
-    print('outside virtualenv or venv')
