--- NLP_LLM/requirements.txt
@@ -1,7 +1,0 @@
-PyQt5
-nltk
-numpy
-networkx
-torch
-transformers
-scipy
