+++ app/checkVenv.py
@@ -1,0 +1,9 @@
+import sys
+
+def is_venv():
+    return (hasattr(sys, 'real_prefix') or
+            (hasattr(sys, 'base_prefix') and sys.base_prefix != sys.prefix))
+if is_venv():
+    print('inside virtualenv or venv')
+else:
+    print('outside virtualenv or venv')
